refactor: reduce RangeField Implementations, break large files
